refactor app init process to use new auth and db code
