+++ package.json
@@ -18,8 +18,6 @@
     "express-validator": "^5.3.1",
     "globby": "^9.2.0",
     "passport": "^0.4.0",
-    "passport-jwt": "^4.0.0",
-    "passport-oauth2-jwt-bearer": "^0.2.0",
     "pg": "^7.11.0",
     "pg-hstore": "^2.3.3",
     "postgrator": "^3.10.1",
@@ -46,7 +44,9 @@
     "husky": "^2.3.0",
     "jest": "^24.8.0",
     "lint-staged": "^8.1.7",
+    "oauth2orize": "^1.11.0",
     "opener": "^1.5.1",
+    "passport-http-bearer": "^1.0.1",
     "prettier": "^1.17.1",
     "prettier-eslint": "^8.8.2",
     "supertest": "^4.0.2"
