+++ src/modules/auth.js
@@ -1,0 +1,36 @@
+import passport from 'passport';
+import BearerStrategy from 'passport-http-bearer';
+
+import models from './database';
+import loggers from './logging';
+
+const log = loggers('auth');
+
+passport.use(
+  new BearerStrategy(async (token, done) => {
+    const { UserToken } = models;
+
+    try {
+      const result = await UserToken.findAll({
+        where: {
+          token
+        }
+      });
+
+      if (!Array.isArray(result) || result.length === 0) {
+        return done(new Error('Authentication failed'));
+      }
+
+      done(null, result, { scope: 'all' });
+    } catch (error) {
+      log.error(error.message);
+      log.error(error.stack);
+      done(error);
+    }
+  })
+);
+
+export default app => {
+  app.use(passport.initialize());
+  app.use(passport.authenticate('bearer', { session: false }));
+};
